requirements.txt
fastapi[standard]
pydantic
google-auth
google-auth-httplib2
google-api-python-client
requests
google-generativeai